+++ dev-server.js
@@ -11,6 +11,7 @@ app.use("/css", express.static(path.join(__dirname, MAIN_PATH + '/css')));
 app.use("/js", express.static(path.join(__dirname, MAIN_PATH + '/js')));
 app.use("/templ", express.static(path.join(__dirname, MAIN_PATH + '/templ')));
 app.use("/locale", express.static(path.join(__dirname, MAIN_PATH + '/locale')));
+app.use("/img", express.static(path.join(__dirname, MAIN_PATH + '/img')));
 app.use("/dist", express.static(path.join(__dirname, 'dist')));
 
 // test api
@@ -20,6 +21,11 @@ app.get("/api/home/", function(req, res) {
 		value: 42
 	});
 });
+
+// anonymizer
+app.get('/anonymizer', function(req, res) {
+	res.sendfile(MAIN_PATH + '/anonymizer.html');
+}); 
 
 // default
 app.get('/*', function(req, res) {