+++ src/q.js
@@ -166,6 +166,64 @@ onix.factory("$q", function() {
 	
 	return {
 		/**
+		 * Inner method for chaining promises.
+		 * 
+		 * @param  {Object[]} opts
+		 * @param  {String|Function} opts.method Function or method name inside scope
+		 * @param  {Object} opts.scope Scope for method function
+		 * @param  {Array} opts.args Additional arguments for function
+		 * @param  {$q} promise Done promise $q
+		 * @param  {Array} outArray Array for output from all executed promises
+		 * @private
+		 * @member $q
+		 */
+		_chainPromisesInner: function(opts, promise, outArray) {
+			var firstItem = opts.shift();
+
+			if (firstItem) {
+				// string or function itself
+				var fn;
+				var error = false;
+
+				switch (typeof firstItem.method) {
+					case "string":
+						if (!firstItem.scope || !(firstItem.method in firstItem.scope)) {
+							error = true;
+						}
+						else {
+							fn = firstItem.scope[firstItem.method];
+
+							if (typeof fn !== "function") {
+								error = true;
+							}
+						}
+						break;
+					case "function":
+						fn = firstItem.method;
+						break;
+					default:
+						error = true;
+				}
+
+				if (!error) {
+					fn.apply(firstItem.scope || fn, firstItem.args || []).then(function(data) {
+						outArray.push(data);
+						this._chainPromisesInner(opts, promise, outArray);
+					}.bind(this), function(err) {
+						outArray.push(err);
+						this._chainPromisesInner(opts, promise, outArray);
+					}.bind(this));
+				}
+				else {
+					promise.resolve(outArray);
+				}
+			}
+			else {
+				promise.resolve(outArray);
+			}
+		},
+
+		/**
 		 * Resolve all promises in the array.
 		 *
 		 * @param {$q[]} promises
@@ -215,6 +273,24 @@ onix.factory("$q", function() {
 		 */
 		isPromise: function(obj) {
 			return obj instanceof $q;
+		},
+
+		/**
+		 * Chaining multiple methods with promises, returns promise.
+		 * 
+		 * @param  {Object[]} opts
+		 * @param  {String|Function} opts.method Function or method name inside scope
+		 * @param  {Object} opts.scope Scope for method function
+		 * @param  {Array} opts.args Additional arguments for function
+		 * @return {$q}
+		 * @member $q
+		 */
+		chainPromises: function(opts) {
+			var promise = this.defer();
+
+			this._chainPromisesInner(opts, promise, []);
+
+			return promise;
 		}
 	};
 });